requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn[standard]
google-generativeai
bs4
langchain-google-genai
arxiv
cassio
langchainhub
semanticscholar
duckduckgo-search
google-search-results
xmltodict
langchain_experimental
pydantic
crewai
crewai_tools
Flask
tavily-python
python_utils
Flask-Cors
